requirements.txt
Flask
SpeechRecognition
PyAudio
openai
python-dotenv